# ticket_management_system/swagger_specs/swagger.yml
swagger: '2.0'
info:
  title: Flight Management System API
  description: RESTful API for managing flights and bookings with scoped user-ID tokens
  version: 1.0.0
basePath: /api
host: localhost:8080
schemes:
- http
- https
consumes:
- application/json
produces:
- application/json
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: >-
      JWT Authorization header using the Bearer scheme.
      Example: "Authorization: Bearer {token}"
parameters:
  PageParam:
    name: page
    in: query
    type: integer
    default: 1
    description: Page number (1-indexed)
  PerPageParam:
    name: per_page
    in: query
    type: integer
    default: 10
    description: Number of items per page (max 100)
...
  FlightIdParam:
    name: flight_id
    in: path
    required: true
    type: string
    format: uuid
    description: Flight identifier
definitions:
...
  ValidationError:
    type: object
    properties:
      error:
        type: string
        example: Bad Request
      message:
        type: string
        example: Validation failed
      errors:
        type: object
        additionalProperties: true
        example:
          email:
          - Not a valid email address
          password:
          - Password must be at least 6 characters
    required:
    - error
    - message
  AuthError:
    type: object
    properties:
      error:
        type: string
        example: Unauthorized
      message:
        type: string
        example: Invalid or missing authentication token
    required:
    - error
    - message
  NotFoundError:
    type: object
    properties:
      error:
        type: string
        example: Not Found
      message:
        type: string
        example: Resource not found
    required:
    - error
    - message
  ConflictError:
    type: object
    properties:
      error:
        type: string
        example: Conflict
      message:
        type: string
        example: Resource already exists or operation cannot be performed
    required:
    - error
    - message
  ForbiddenError:
    type: object
    properties:
      error:
        type: string
        example: Forbidden
      message:
        type: string
        example: You do not have permission to access this resource
    required:
    - error
    - message
...
  Flight:
    type: object
    properties:
      id:
        type: string
        format: uuid
      flight_code:
        type: string
      origin_airport:
        type: string
      destination_airport:
        type: string
      departure_time:
        type: string
        format: date-time
      arrival_time:
        type: string
        format: date-time
      base_price:
        type: string
      status:
        type: string
        enum:
        - active
        - inactive
        - started
        - en_route
        - landed
        - cancelled
        - delayed
      created_at:
        type: string
        format: date-time
      updated_at:
        type: string
        format: date-time
  Pagination:
    type: object
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total_pages:
        type: integer
      total_items:
        type: integer
      has_next:
        type: boolean
      has_prev:
        type: boolean
      next_page:
        type: integer
        nullable: true
      prev_page:
        type: integer
        nullable: true
  FlightListResponse:
    type: object
    properties:
      flights:
        type: array
        items:
          $ref: '#/definitions/Flight'
      pagination:
        $ref: '#/definitions/Pagination'
...
  AirportListResponse:
    type: object
    properties:
      airports:
        type: array
        items:
          type: string
      count:
        type: integer
...
paths:
...
  /flights/airports:
    get:
      summary: List all available airports
      description: Get a list of all distinct airports available in the system.
      operationId: listAirports
      tags:
      - Flights
      security:
      - Bearer: []
      responses:
        200:
          description: List of airports
          schema:
            $ref: '#/definitions/AirportListResponse'
        401:
          description: Unauthorized
          schema:
            $ref: '#/definitions/AuthError'
          examples:
            application/json: *auth_error_example
  /flights/search:
    get:
      summary: Search flights
      description: Search for flights with optional filters. Returns paginated results.
      operationId: searchFlights
      tags:
      - Flights
      security:
      - Bearer: []
      parameters:
      - name: status
        in: query
        type: string
        description: Filter by flight status (defaults to active only)
        enum:
        - all
        - active
        - inactive
        - started
        - en_route
        - landed
        - cancelled
        - delayed
      - name: origin_airport
        in: query
        type: string
        description: Filter by origin airport
      - name: destination_airport
        in: query
        type: string
        description: Filter by destination airport
      - name: departure_date
        in: query
        type: string
        format: date
        description: Filter by departure date (YYYY-MM-DD)
      - name: arrival_date
        in: query
        type: string
        format: date
        description: Filter by arrival date (YYYY-MM-DD)
      - name: sort_by
        in: query
        type: string
        description: Sort field for results
        enum:
        - departure_time
        - arrival_time
        - base_price
        default: departure_time
      - name: sort_order
        in: query
        type: string
        description: Sort direction
        enum:
        - asc
        - desc
        default: asc
      - $ref: '#/parameters/PageParam'
      - $ref: '#/parameters/PerPageParam'
      responses:
        200:
          description: Paginated list of flights
          schema:
            $ref: '#/definitions/FlightListResponse'
        400:
          description: Validation error
          schema:
            $ref: '#/definitions/ValidationError'
          examples:
            application/json: *validation_error_example
        401:
          description: Unauthorized
          schema:
            $ref: '#/definitions/AuthError'
          examples:
            application/json: *auth_error_example
  /flights:
    post:
      summary: Add a new flight (Admin only)
      description: Create a new flight in the system. Only administrators can add flights.
      operationId: addFlight
      tags:
      - Flights
      security:
      - Bearer: []
      parameters:
      - in: body
        name: body
        required: true
        schema:
          type: object
          required:
          - flight_code
          - origin_airport
          - destination_airport
          - departure_time
          - arrival_time
          - base_price
          properties:
            flight_code:
              type: string
              minLength: 3
              maxLength: 8
              example: AA1234
            origin_airport:
              type: string
              example: JFK
            destination_airport:
              type: string
              example: LAX
            departure_time:
              type: string
              format: date-time
              example: '2026-12-25 10:30:00'
            arrival_time:
              type: string
              format: date-time
              example: '2026-12-25 14:45:00'
            base_price:
              type: number
              minimum: 0
              example: 299.99
      responses:
        201:
          description: Flight created successfully
          schema:
            type: object
            properties:
              message:
                type: string
              flight:
                $ref: '#/definitions/Flight'
        400:
          description: Validation error
          schema:
            $ref: '#/definitions/ValidationError'
          examples:
            application/json: *validation_error_example
        401:
          description: Unauthorized
          schema:
            $ref: '#/definitions/AuthError'
          examples:
            application/json: *auth_error_example
        403:
          description: Forbidden - Admin privileges required
          schema:
            $ref: '#/definitions/ForbiddenError'
          examples:
            application/json: *forbidden_error_example
        409:
          description: Flight code already exists
          schema:
            $ref: '#/definitions/ConflictError'
          examples:
            application/json: *conflict_error_example
  /flights/{flight_id}:
    get:
      summary: Get flight by ID
      description: Retrieve details of a specific flight by its UUID.
      operationId: getFlightById
      tags:
      - Flights
      security:
      - Bearer: []
      parameters:
      - $ref: '#/parameters/FlightIdParam'
      responses:
        200:
          description: Flight details
          schema:
            type: object
            properties:
              flight:
                $ref: '#/definitions/Flight'
        401:
          description: Unauthorized
          schema:
            $ref: '#/definitions/AuthError'
          examples:
            application/json: *auth_error_example
        404:
          description: Flight not found
          schema:
            $ref: '#/definitions/NotFoundError'
          examples:
            application/json: *not_found_error_example
    delete:
      summary: Delete a flight (Admin only)
      description: Delete a flight from the system. Only administrators can delete flights.
      operationId: deleteFlight
      tags:
      - Flights
      security:
      - Bearer: []
      parameters:
      - $ref: '#/parameters/FlightIdParam'
      responses:
        200:
          description: Flight deleted successfully
          schema:
            type: object
            properties:
              message:
                type: string
        401:
          description: Unauthorized
          schema:
            $ref: '#/definitions/AuthError'
          examples:
            application/json: *auth_error_example
        403:
          description: Forbidden - Admin privileges required
          schema:
            $ref: '#/definitions/ForbiddenError'
          examples:
            application/json: *forbidden_error_example
        404:
          description: Flight not found
          schema:
            $ref: '#/definitions/NotFoundError'
          examples:
            application/json: *not_found_error_example
    put:
      summary: Update a flight (Admin only)
      description: Update flight details. Only administrators can update flights.
      operationId: updateFlight
      tags:
      - Flights
      security:
      - Bearer: []
      parameters:
      - $ref: '#/parameters/FlightIdParam'
      - in: body
        name: body
        required: true
        schema:
          type: object
          properties:
            origin_airport:
              type: string
            destination_airport:
              type: string
            departure_time:
              type: string
              format: date-time
            arrival_time:
              type: string
              format: date-time
            base_price:
              type: number
            status:
              type: string
              enum:
              - active
              - inactive
              - started
              - en_route
              - landed
              - cancelled
              - delayed
      responses:
        200:
          description: Flight updated successfully
          schema:
            type: object
            properties:
              message:
                type: string
              flight:
                $ref: '#/definitions/Flight'
        400:
          description: Validation error
          schema:
            $ref: '#/definitions/ValidationError'
          examples:
            application/json: *validation_error_example
        401:
          description: Unauthorized
          schema:
            $ref: '#/definitions/AuthError'
          examples:
            application/json: *auth_error_example
        403:
          description: Forbidden - Admin privileges required
          schema:
            $ref: '#/definitions/ForbiddenError'
          examples:
            application/json: *forbidden_error_example
        404:
          description: Flight not found
          schema:
            $ref: '#/definitions/NotFoundError'
          examples:
            application/json: *not_found_error_example
...